@@ -10,18 +10,14 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "firebase": "^9.0.0-beta.2",
-    "vue": "^3.0.0",
-    "vue-router": "^4.0.0-0"
+    "firebase": "^9.1.3",
+    "vue": "^3.2.0",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2",
+    "vuex-persistedstate": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0"
+    "@vue/cli-service": "^4.5.14",
+    "eslint": "^8.0.1"
   }
 }
