@@ -1,72 +1,27 @@
-import { initializeApp } from "firebase/app"
-import { getFirestore } from "firebase/firestore";
-
 import { config } from '@/firebaseApiKey';
+import { initializeApp } from "firebase/app"
+import { getAuth, onAuthStateChanged } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+import { getStorage, ref } from "firebase/storage";
+
+import store from '@/store'
+
 // import { ref, onUnmounted } from 'vue';
 
+// Init Firebase
 export const firebaseApp = initializeApp(config);
 
+// Firebase auth
+const auth = getAuth();
+onAuthStateChanged(auth, (user) => {
+  store.dispatch("fetchUser", user);
+});
+
+// Firebase cloud firestore
 export const db = getFirestore();
 
 
+// Firebase storage
+const storage = getStorage(firebaseApp);
 
-
-// export const createDocument = (collection, document) => {
-//     return db.collection(collection).add(document);
-// }
-
-// export const getDocumentById = async (collection, id) => {
-//     const employee = await  db.collection(collection).doc(id).get();
-//     return employee.exists ? employee.data() : null;
-// }
-
-// export const updateDocument = (collection, id, document) => {
-//     return db.collection(collection).doc(id).update(document);
-// }
-
-// export const deleteDocument = (collection, id) => {
-//     return db.collection(collection).doc(id).delete();
-// }
-
-// export const getAllDocuments = (collection) => {
-//     const documents = ref([]);
-//     const close = db.collection(collection).onSnapshot(snapshot => {
-//         documents.value = snapshot.docs.map(doc => ({ id : doc.id, ... doc.data() }))
-//     });
-//     onUnmounted(close);
-//     return documents;
-// }
-
-
-
-
-
-
-// const employeeCollection = db.collection("employee");
-
-// export const createEmployee = employee => {
-//     return employeeCollection.add(employee);
-// }
-
-// export const getEmployee = async id => {
-//     const employee = await employeeCollection.doc(id).get();
-//     return employee.exists ? employee.data() : null;
-// }
-
-// export const updateEmployee = (id, employee) => {
-//     return employeeCollection.doc(id).update(employee);
-// }
-
-// export const deleteEmployee = id => {
-//     return employeeCollection.doc(id).delete();
-// }
-
-// export const useLoadEmployees = () => {
-//     const employees = ref([]);
-//     const close = employeeCollection.onSnapshot(snapshot => {
-//         employees.value = snapshot.docs.map(doc => ({ id : doc.id, ... doc.data() }))
-//     });
-//     onUnmounted(close);
-//     return employees;
-// }
-
+// export const storageRef = ref(storage);
